pyproject.toml
[project]
name = "safe-jax"
version = "0.1.0"
description = "SAFE: Sharpness-Aware Flatness Enforcement for Sparse Network Training"
requires-python = ">=3.8"
dependencies = [
    "absl-py==1.0.0",
    "chex==0.1.6",
    "clu==0.0.9",
    "datasets==2.10.1",
    "flax==0.6.5",
    "jax==0.4.4",
    "jaxlib==0.4.4+cuda11.cudnn82",
    "ml-collections==0.1.0",
    "numpy==1.22.0",
    "nvidia-cudnn-cu11==8.6.0.163",
    "optax==0.1.3",
    "orbax==0.1.2",
    "scikit-learn==1.0.0",
    "tensorflow==2.9.3",
    "tensorflow-datasets==4.4.0",
    "torch==1.12.0",
    "tensorflow-io-gcs-filesystem<0.35",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.uv]
find-links = ["https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"]
